+++ src/server/cards/base/RoboticWorkforce.ts
@@ -1,18 +1,12 @@
-import {IProjectCard} from '../IProjectCard';
 import {Tag} from '../../../common/cards/Tag';
-import {Card} from '../Card';
+import {RoboticWorkforceBase} from './RoboticWorkforceBase';
 import {CardType} from '../../../common/cards/CardType';
-import {IPlayer} from '../../IPlayer';
-import {SelectCard} from '../../inputs/SelectCard';
 import {CardName} from '../../../common/cards/CardName';
-import {ICard} from '../ICard';
 import {CardRenderer} from '../render/CardRenderer';
 import {Size} from '../../../common/cards/render/Size';
 import {played} from '../Options';
-import {Behavior} from '../../behavior/Behavior';
-import {getBehaviorExecutor} from '../../behavior/BehaviorExecutor';
 
-export class RoboticWorkforce extends Card implements IProjectCard {
+export class RoboticWorkforce extends RoboticWorkforceBase {
   constructor() {
     super({
       type: CardType.AUTOMATED,
@@ -30,68 +24,11 @@ export class RoboticWorkforce extends Card implements IProjectCard {
     });
   }
 
-  /**
-   * Returns a copy of behavior with just `production` and `decreaseAnyProduction` fields.
-   */
-  private productionBehavior(behavior: Behavior): Behavior {
-    const filtered: Behavior = {};
-    if (behavior.production !== undefined) {
-      filtered.production = behavior.production;
-    }
-    if (behavior.decreaseAnyProduction !== undefined) {
-      filtered.decreaseAnyProduction = behavior.decreaseAnyProduction;
-    }
-    return filtered;
+  public override selectCardText(): string {
+    return 'Select builder card to copy';
   }
 
-  private isCardApplicable(card: ICard, player: IPlayer): boolean {
-    if (!card.tags.includes(Tag.BUILDING) && !card.tags.includes(Tag.WILD)) {
-      return false;
-    }
-    if (card.name === CardName.SPECIALIZED_SETTLEMENT) {
-      return player.production.energy >= 1;
-    }
-
-    if (card.produce !== undefined) {
-      return true;
-    }
-
-    if (card.behavior !== undefined) {
-      const productionBehavior = this.productionBehavior(card.behavior);
-      if (Object.keys(productionBehavior).length > 0) {
-        return getBehaviorExecutor().canExecute(productionBehavior, player, card);
-      }
-    }
-
-    // Card has no production box.
-    return false;
-  }
-
-  private getAvailableCards(player: IPlayer): Array<ICard> {
-    return player.tableau.filter((card) => this.isCardApplicable(card, player));
-  }
-
-  public override bespokeCanPlay(player: IPlayer): boolean {
-    return this.getAvailableCards(player).length > 0;
-  }
-
-  public override bespokePlay(player: IPlayer) {
-    const availableCards = this.getAvailableCards(player);
-
-    if (availableCards.length === 0) {
-      return undefined;
-    }
-
-    return new SelectCard('Select builder card to copy', 'Copy', availableCards, ([card]) => {
-      player.game.log('${0} copied ${1} production with ${2}', (b) =>
-        b.player(player).card(card).card(this));
-
-      if (card.produce) {
-        card.produce(player);
-      } else if (card.behavior !== undefined) {
-        getBehaviorExecutor().execute(this.productionBehavior(card.behavior), player, card);
-      }
-      return undefined;
-    });
+  public override count(): number {
+    return 1;
   }
 }
